+++ @angular-generic-table/core/interfaces/gt-event.ts
@@ -1,0 +1,4 @@
+export interface GtEvent {
+  name: string;
+  value: any;
+}