+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Release History
 ### Added
 - Highlight pipe for use in custom components #196
 
+### Fixed
+- Page change event only emitted when page has actually changed #200
+
 # [4.11.1] - 2018-01-19
 ### Fixed
 - Condition check in gt render pipe for 0, null, '' and false values #178